~ Version : Debug
Keep testing and developing debug versions for the customer to have a view of what's going on, in a way, that makes it possible to have an understanding of the system itself. Fully done with the design and small changes of how it will end up working. Still in progress. Further releases will appear.