[Frustration Update] ~ Version : 0.10.4
Slowly fixing and creating new problems. Trying to fix the situation, for when a User is logging onto the website, everything is correct. When the website is reloaed, wrong information is loaded onto the site, and there is no tracking or information, to how this information is gathered and displayed wrong. No idea how to proceed from here. Good luck.