[Website] - v1.5.0 - .NET 5.0 Upgrade
Contracts are now fully functional, updated a new way of creating contracts for a company, it's now possible to show contracts in pdf format on the site, without inconsistency. Haven't checked for further errors or bugs. They will come eventually i think.

Trying to play around with the JS Interlop and see if the site can be a bit quicker and smarter with the way it's being used right now.