Attempted to fix.
Been playing around with a fix for the creation of contracts in the database. Figured out the issues, whenever a user / company doesn't exists in the DB. It just passes through the If statements and goes directly to the catch option. Have to find a new way to error handle.