Creating and Updating data.
This version has been enchanced with changes for the Vacant Jobs & Job Adverts, data is being checked for if it already exists, but also created if it doesn't. But for job adverts, data is being updated if it already exists, where for vacant jobs, the iteration of the data list is just continued on, if it already exists. Created a method for getting data for vacant jobs, and modified a few of the methods in the manager.