A Magical Update
Removed pre-created templates from the framework.
Added the remaining pages to be worked with, but removed the pre-created pages that comes with the framework. Fixed a namespace error in the components, that were calling the models.

Implemented role authorization in the sidebar menu for when users are logged in, added the authorize attribute to all pages, but out-commented so it's not in use before the login is implemented.

Removed services from the startup that were used by the framework itself for demo purposes.