A little update.
Have tested the project that is it able to launch without breaking. It's just crashing on the pages since it cannot find a certain amount of procedures, since they haven't been made or the DB hasn't been updated on the environment, otherwise, tried implementing the old code into the new design for the pages.