Small Change
Tried implementing a provider to clear messages after a specific internal, is buggy.

Implementing the functionality for the my profile page, has to have some new methods regarding the update of the user, the live update of the consultant areas of the user, which is not part of the main editform itself. But updates when the user adds an area to their consultant areas.