More User Stuff
Finished the User Profile Dropdown menu with functionality, it updates whenever the user data is updated. The Dropdown just needs some styling. Been trying to fix the forms on the My Profile page, it's being buggy with the edit context.

Changed a few things about the Concrete User and it's Interface. Whilst also updated the IUserService to provide less methods, but more logic inside the methods itself, instead of just providing the ability to give access to the repository.