Quickiiie Fix
Modified the Job Advert Modals to behave as needed, can successfully create a job advert now.

A bug is found when trying to edit a job advert also while trying to delete the job advert.

Fixed the job advert model to display error messages if anything in the form is missing.

Changed the data access to match the database changes, so all data is given correctly.

Added new icons to the control-panel, in order to differentiate between the links.